Speedup of the test-kernel computation
During the scoring-/test-phase the kernels between training and candidates
have to be calculated. In the former implementation even the kernel between
all the candidates was computed. However, only the diagonal of the that
kernel-matrix is needed.

Now only the diagonal elements rather than the full kernel-matrix are
computed. This leeds to an huge speedup.